use env file

+++ upload.sh
@@ -1,6 +1,14 @@
 #!/usr/bin/env bash
 
-server="root@100.106.77.84"
+source .env
+
+# Error if no HA_IP is set
+if [ -z "$HA_IP" ]; then
+  echo "Please set HA_IP in .env file"
+  exit 1
+fi
+
+server="root@$HA_IP"
 
 ssh "$server" "rm -rv /config/automations /config/scenes"
